requirements.txt
accelerate==1.2.1
aenum==3.1.15
aioboto3==13.3.0
aiobotocore==2.16.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioitertools==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
anytree==2.12.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==24.3.0
autograd==1.7.0
beartype==0.18.5
boto3==1.35.81
botocore==1.35.81
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
gensim==4.3.3
graspologic==3.4.1
graspologic-native==1.2.1
greenlet==3.1.1
gremlinpython==3.7.3
grpcio==1.67.1
h11==0.14.0
hnswlib==0.8.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.27.0
hyppo==0.4.0
idna==3.10
ipython==8.31.0
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpickle==4.0.1
kiwisolver==1.4.7
lightrag-hku==1.0.7
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
milvus-lite==2.4.10
mpmath==1.3.0
multidict==6.1.0
nano-vectordb==0.0.4.3
neo4j==5.27.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ollama==0.4.4
openai==1.58.1
oracledb==2.5.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.0.0
POT==0.9.5
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.4
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pymilvus==2.5.1
pymongo==4.10.1
PyMySQL==1.1.1
pynndescent==0.5.13
pyparsing==3.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
s3transfer==0.10.4
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.12.0
seaborn==0.13.2
sentence-transformers==3.3.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
umap-learn==0.5.7
urllib3==2.3.0
wcwidth==0.2.13
wrapt==1.17.0
xxhash==3.5.0
yarl==1.18.3
